move navigateWhenEnableFeature
